Derivative-Free Optimization

Derivative-Free optimization algorithms. These algorithms do not require gradient information. More importantly, they can be used to solve non-smooth optimization problems. They can also handle box constraints on parameters.

Details

Package: 	 dfoptim

Type: 	 Package

Version: 	 2016.7-1

Date: 	 2016-07-08

License: 	 GPL-2 or greater

LazyLoad: 	 yes

Derivative-Free optimization algorithms. These algorithms do not require gradient information. More importantly, they can be used to solve non-smooth optimization problems. These algorithms were translated from the Matlab code of Prof. C.T. Kelley, given in his book "Iterative methods for optimization". However, there are some non-trivial modifications of the algorithm.

Currently, the Nelder-Mead and Hooke-Jeeves algorithms is implemented. In future, more derivative-free algorithms may be added.

References

C.T. Kelley (1999), Iterative Methods for Optimization, SIAM.